Fix (hopefully) a bug with the screen jittering around
It seems like some terminal emulators (xterm, whatever windows uses by
default now) don't like our trying to write to the whole screen, and as
a result the whole screen content jitters around as if we were trying to
write to more rows than we have. A somewhat hacky fix is to simply act
like we have slightly fewer lines than we do.